feat: 경주 레이스 배팅 금액 범위 조정

+++ src/main/java/woowa/gamble/controller/RacingController.java
@@ -50,7 +50,7 @@ public class RacingController {
     public String playRace(@RequestParam int multiplier,
                            @RequestParam int carCount,
                            @RequestParam String selectedCar,
-                           @RequestParam int betAmount,
+                           @RequestParam Long betAmount,
                            Model model, HttpSession session) {
         Long userId = (Long) session.getAttribute("myUserId");
         if (userId == null) return "redirect:/";
